Convert drfti1 to a safe version
I am still not satisfied with this, but at least it can be a decent
starting point for improvements. My points are:
1. Variable names are still unintelligible, mainly because I do not have
a clear vision of the algorithm. Maybe someone better than me can be
able to grasp the meaning of each variable and to git them a decent
name.
2. The first loop is still unidiomatic. Everything I thought involves an
additional check to exit the outer loop, and at the current state (no
benchmarks) I am not sure about the best approch.

Nevertheless, cast checks that were not present in the original code
should be helpful to find strange edge cases, and they should have a
negligible cost due to the cpu branch predictor.