Add original version of fft and update tests
Instead of trying to find a similar FFT algorithm that could help
simulating the current behaviour, it is easier to use the initial
version of the rust fft files with a minimal set of changes.

The tests have been updated in order to compare the current
implementation with the original one. `drftf1` test has been modified to
have 100% coverage.